Refactored user management components to be accessible by BBMRI admin role
