+++ podium-gateway/src/main/webapp/app/admin/index.ts
@@ -34,11 +34,5 @@ export * from './metrics/metrics.component';
 export * from './metrics/metrics-modal.component';
 export * from './metrics/metrics.service';
 export * from './metrics/metrics.route';
-export * from './user-management/user-management-dialog.component';
-export * from './user-management/user-management-delete-dialog.component';
-export * from './user-management/user-management-detail.component';
-export * from './user-management/user-management-unlock-dialog.component';
-export * from './user-management/user-management.component';
-export * from './user-management/user-management.route';
-export * from './user-management/user-modal.service';
+
 export * from './admin.route';
