+++ podium-gateway/src/main/webapp/app/entities/user-management/index.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2017. The Hyve and respective contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * See the file LICENSE in the root of this repository.
+ *
+ */
+export * from './user-management-dialog.component';
+export * from './user-management-delete-dialog.component';
+export * from './user-management-detail.component';
+export * from './user-management-unlock-dialog.component';
+export * from './user-management.component';
+export * from './user-management.route';
+export * from './user-modal.service';
