+++ podium-gateway/src/main/webapp/app/entities/user-management/user-management-delete-dialog.component.html
