+++ podium-gateway/src/main/webapp/app/entities/user-management/user-management-delete-dialog.component.ts
@@ -12,8 +12,8 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NgbActiveModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { EventManager, JhiLanguageService } from 'ng-jhipster';
-
-import { User, UserService } from '../../shared';
+import { UserService } from '../../shared/user/user.service';
+import { User } from '../../shared/user/user.model';
 import { UserModalService } from './user-modal.service';
 
 @Component({
