+++ podium-gateway/src/main/webapp/app/entities/user-management/user-management-detail.component.ts
