+++ podium-gateway/src/main/webapp/app/entities/user-management/user-management-unlock-dialog.component.ts
@@ -13,7 +13,8 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { NgbActiveModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { EventManager, JhiLanguageService } from 'ng-jhipster';
 
-import { User, UserService } from '../../shared';
+import { UserService } from '../../shared/user/user.service';
+import { User } from '../../shared/user/user.model';
 import { UserModalService } from './user-modal.service';
 
 @Component({
