+++ podium-gateway/src/main/webapp/app/entities/user-management/user-management.route.ts
@@ -11,7 +11,7 @@
 import { Injectable } from '@angular/core';
 import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot, Routes, CanActivate } from '@angular/router';
 
-import { UserRouteAccessService } from '../../shared';
+// import { UserRouteAccessService } from '../.';
 import { PaginationUtil } from 'ng-jhipster';
 
 import { UserMgmtComponent } from './user-management.component';
@@ -19,8 +19,9 @@ import { UserMgmtDetailComponent } from './user-management-detail.component';
 import { UserDialogComponent } from './user-management-dialog.component';
 import { UserDeleteDialogComponent } from './user-management-delete-dialog.component';
 
-import { Principal } from '../../shared';
+// import { Principal } from '../.';
 import { UserUnlockDialogComponent } from './user-management-unlock-dialog.component';
+import { Principal } from '../../shared/auth/principal.service';
 
 
 @Injectable()
